I swear I am done with refactoring
