+++ client/components/interface/Main/Main.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-unused-vars */
 import React, { useState } from "react";
 import { StyledMain } from "./Main.styled";
-import Routes from "../../routes/Routes";
-import Nav from "../interface/Nav/Nav";
-import Burger from "../utils/Burger/Burger";
-import Menu from "../views/Menu/Menu";
-import Footer from "../interface/Footer/Footer";
+import Routes from "../../../routes/Routes";
+import Nav from "../Nav/Nav";
+import Burger from "../../utils/Burger/Burger";
+import Menu from "../../views/Menu/Menu";
+import Footer from "../Footer/Footer";
 import { BrowserRouter as Router } from "react-router-dom";
 
 const Main = () => {
